feat: extract part autocomplete component
